+++ Runtime/ContainerManager/Interface/ISceneContainerItem.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine.SceneManagement;
+
+namespace Cobilas.Unity.Management.Container {
+    public interface ISceneContainerItem {
+        void sceneUnloaded(Scene scene);
+        void sceneLoaded(Scene scene, LoadSceneMode mode);
+    }
+}