+++ Runtime/InitializeMethod/Old_InitializeMethod/Initialize/Enumerators/CRIOLMType.cs
@@ -1,6 +1,0 @@
-namespace Cobilas.Unity.Management.Runtime {
-    public enum CRIOLMType {
-        AfterSceneLoad = 0,
-        BeforeSceneLoad = 1
-    }
-}